Updated build options to support stable202102 builds
This removes the option for an autoboot image and instead needs
to be built specifically for sd for flexspi_nor, with the default
being for sd.

This build also defaults to the X86EMU support being disabled by
default and can be enabled with the X86EMU=1 environment option.
This is disabled by default as the BBR spec states that emulation
for drivers is not allowed, however we will leave it as a developer
option.

The AMDGOP driver is enabled by default but can be disabled with the
AMDGOP=0 build option.

XMP profile can be enabled by setting the option XMP_PROFILE= 1 or 2
and setting the appropriate DDR_SPEED to the profile target memory
speed.  This is experimental and disabled by default.

Signed-off-by: Jon Nettleton <jon@solid-run.com>