Make build script more robust and add docker builds
If doing various builds from the same directory TF-A will
gladly create binaries without the drivers reflected in the
build type.  Therefore we remove the TF-A build directory
on each run since it is not very large.

This also adds the ability to build images in a docker or podman
container.

image names have changed, adding both the targeted boot medium
and the main lx2160a_uefi git hash to filename so they can
be tracked better.

Signed-off-by: Jon Nettleton <jon@solid-run.com>